+++ src/Overlay.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import BaseOverlay, {
@@ -8,7 +8,6 @@ import BaseOverlay, {
 } from '@restart/ui/Overlay';
 import { State } from '@restart/ui/usePopper';
 import { componentOrElement, elementType } from 'prop-types-extra';
-import useCallbackRef from '@restart/hooks/useCallbackRef';
 import useEventCallback from '@restart/hooks/useEventCallback';
 import useIsomorphicEffect from '@restart/hooks/useIsomorphicEffect';
 import useMergedRefs from '@restart/hooks/useMergedRefs';
@@ -28,6 +27,8 @@ export interface OverlayInjectedProps {
   show: boolean;
   placement: Placement | undefined;
   popper: PopperRef;
+  hasDoneInitialMeasure?: boolean;
+
   [prop: string]: any;
 }
 
@@ -140,13 +141,6 @@ const propTypes = {
   ]),
 };
 
-const defaultProps: Partial<OverlayProps> = {
-  transition: Fade,
-  rootClose: false,
-  show: false,
-  placement: 'top',
-};
-
 function wrapRefs(props, arrowProps) {
   const { ref } = props;
   const { ref: aRef } = arrowProps;
@@ -158,11 +152,21 @@ function wrapRefs(props, arrowProps) {
 
 const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
   (
-    { children: overlay, transition, popperConfig = {}, ...outerProps },
+    {
+      children: overlay,
+      transition = Fade,
+      popperConfig = {},
+      rootClose = false,
+      placement = 'top',
+      show: outerShow = false,
+      ...outerProps
+    },
     outerRef,
   ) => {
     const popperRef = useRef<Partial<PopperRef>>({});
-    const [firstRenderedState, setFirstRenderedState] = useCallbackRef<State>();
+    const [firstRenderedState, setFirstRenderedState] = useState<State | null>(
+      null,
+    );
     const [ref, modifiers] = useOverlayOffset(outerProps.offset);
     const mergedRef = useMergedRefs(outerRef, ref);
 
@@ -180,6 +184,12 @@ const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
       }
     }, [firstRenderedState]);
 
+    useEffect(() => {
+      if (!outerShow) {
+        setFirstRenderedState(null);
+      }
+    }, [outerShow]);
+
     return (
       <BaseOverlay
         {...outerProps}
@@ -190,6 +200,9 @@ const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
           onFirstUpdate: handleFirstUpdate,
         }}
         transition={actualTransition}
+        rootClose={rootClose}
+        placement={placement}
+        show={outerShow}
       >
         {(overlayProps, { arrowProps, popper: popperObj, show }) => {
           wrapRefs(overlayProps, arrowProps);
@@ -201,7 +214,10 @@ const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
             placement: updatedPlacement,
             outOfBoundaries:
               popperObj?.state?.modifiersData.hide?.isReferenceHidden || false,
+            strategy: popperConfig.strategy,
           });
+
+          const hasDoneInitialMeasure = !!firstRenderedState;
 
           if (typeof overlay === 'function')
             return overlay({
@@ -211,6 +227,7 @@ const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
               ...(!transition && show && { className: 'show' }),
               popper,
               arrowProps,
+              hasDoneInitialMeasure,
             });
 
           return React.cloneElement(overlay as React.ReactElement, {
@@ -218,6 +235,7 @@ const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
             placement: updatedPlacement,
             arrowProps,
             popper,
+            hasDoneInitialMeasure,
             className: classNames(
               (overlay as React.ReactElement).props.className,
               !transition && show && 'show',
@@ -235,6 +253,5 @@ const Overlay = React.forwardRef<HTMLElement, OverlayProps>(
 
 Overlay.displayName = 'Overlay';
 Overlay.propTypes = propTypes;
-Overlay.defaultProps = defaultProps;
 
 export default Overlay;