+++ src/OverlayTrigger.tsx
@@ -164,13 +164,8 @@ const propTypes = {
   ]),
 };
 
-const defaultProps = {
-  defaultShow: false,
-  trigger: ['hover', 'focus'],
-};
-
 function OverlayTrigger({
-  trigger,
+  trigger = ['hover', 'focus'],
   overlay,
   children,
   popperConfig = {},
@@ -315,6 +310,5 @@ function OverlayTrigger({
 }
 
 OverlayTrigger.propTypes = propTypes;
-OverlayTrigger.defaultProps = defaultProps;
 
 export default OverlayTrigger;