+++ src/Popover.tsx
@@ -7,6 +7,7 @@ import PopoverHeader from './PopoverHeader';
 import PopoverBody from './PopoverBody';
 import { Placement, PopperRef } from './types';
 import { BsPrefixProps, getOverlayDirection } from './helpers';
+import getInitialPopperStyles from './getInitialPopperStyles';
 
 export interface PopoverProps
   extends React.HTMLAttributes<HTMLDivElement>,
@@ -17,6 +18,7 @@ export interface PopoverProps
   body?: boolean;
   popper?: PopperRef;
   show?: boolean;
+  hasDoneInitialMeasure?: boolean;
 }
 
 const propTypes = {
@@ -71,6 +73,11 @@ const propTypes = {
    */
   body: PropTypes.bool,
 
+  /**
+   * Whether or not Popper has done its initial measurement and positioning.
+   */
+  hasDoneInitialMeasure: PropTypes.bool,
+
   /** @private */
   popper: PropTypes.object,
 
@@ -78,22 +85,19 @@ const propTypes = {
   show: PropTypes.bool,
 };
 
-const defaultProps: Partial<PopoverProps> = {
-  placement: 'right',
-};
-
 const Popover = React.forwardRef<HTMLDivElement, PopoverProps>(
   (
     {
       bsPrefix,
-      placement,
+      placement = 'right',
       className,
       style,
       children,
       body,
       arrowProps,
-      popper: _,
-      show: _1,
+      hasDoneInitialMeasure,
+      popper,
+      show,
       ...props
     },
     ref,
@@ -103,11 +107,19 @@ const Popover = React.forwardRef<HTMLDivElement, PopoverProps>(
     const [primaryPlacement] = placement?.split('-') || [];
     const bsDirection = getOverlayDirection(primaryPlacement, isRTL);
 
+    let computedStyle = style;
+    if (show && !hasDoneInitialMeasure) {
+      computedStyle = {
+        ...style,
+        ...getInitialPopperStyles(popper?.strategy),
+      };
+    }
+
     return (
       <div
         ref={ref}
         role="tooltip"
-        style={style}
+        style={computedStyle}
         x-placement={primaryPlacement}
         className={classNames(
           className,
@@ -124,7 +136,6 @@ const Popover = React.forwardRef<HTMLDivElement, PopoverProps>(
 );
 
 Popover.propTypes = propTypes as any;
-Popover.defaultProps = defaultProps;
 
 export default Object.assign(Popover, {
   Header: PopoverHeader,