+++ www/src/examples/Toast/PlacementMulti.js
@@ -12,7 +12,7 @@ function PlacementMultiExample() {
       className="bg-dark position-relative"
       style={{ minHeight: '240px' }}
     >
-      <ToastContainer position="top-end" className="p-3">
+      <ToastContainer position="top-end" className="p-3" style={{ zIndex: 1 }}>
         <Toast>
           <Toast.Header>
             <img