+++ www/src/examples/Toast/Stacking.js
@@ -3,10 +3,12 @@ import Toast from 'react-bootstrap/Toast';
 import ToastContainer from 'react-bootstrap/ToastContainer';
 
 function StackingExample() {
-  const [position, setPosition] = useState('top-start');
-
+  const codeExampleStylingFix = {
+    zIndex: 1,
+    position: 'relative',
+  };
   return (
-    <ToastContainer>
+    <ToastContainer style={codeExampleStylingFix}>
       <Toast>
         <Toast.Header>
           <img src="holder.js/20x20?text=%20" className="rounded me-2" alt="" />